Integracion de interfaz en clase de Garaxe

+++ Boletin 10/Repaso Boletin 10/src/calculoImporte.java
@@ -1,5 +1,5 @@
 public interface calculoImporte {
 
-    int  calculadoraImporte();
+    double  calculadoraImporte();
 
 }//end interface
